alterações para verificação de erro no heroku

+++ src/main/resources/static/js/interceptor/token-inteceptor.js
@@ -10,10 +10,10 @@ aqtApp.factory("tokenInterceptor", function($q, $location){
 		},
 		
 		  'responseError': function (rejection){
-	 /*   	if(rejection.status==401){
+	    	if(rejection.status==401){
 	    		localStorage.clear();
 	    		$location.path("/aqtAdmin");
-	    	}*/
+	    	}
 	    	return rejection;
 	    }
 	};
